requirements.txt
asgiref==3.8.1
Django==3.0.11
djangorestframework==3.11.2
mysqlclient==2.2.4
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
